% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_selected_mutation.R
\name{add_mut_to_pheno}
\alias{add_mut_to_pheno}
\title{Add phenotypic effects of mutation(s) from \code{pheno_mut_effect} to a given
phenotype \code{pheno_wt}.}
\usage{
add_mut_to_pheno(pheno_mut_effect, pheno_wt)
}
\arguments{
\item{pheno_mut_effect}{A matrix of real numbers. Each row correspond
to the phenotypic effects of given mutation in each dimensions (columns).}

\item{pheno_wt}{A vector of real numbers. Phenotypic coordinates of the wild-type.
Its length must be equal to the number of columns of \code{pheno_mut_effect}.}
}
\value{
A matrix of the same dimensions as \code{pheno_rand_mut_effect} with
each row equal to the sum of \code{pheno_wt} with a row of \code{pheno_rand_mut_effect}.
}
\description{
Add phenotypic effects of mutation(s) from \code{pheno_mut_effect} to a given
phenotype \code{pheno_wt}.
}
