% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{assert_genotype_table}
\alias{assert_genotype_table}
\title{Checkmate wrapper for argument checks. See \code{\link[checkmate]{makeAssertion}}
for more informations.}
\usage{
assert_genotype_table(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The matrix to check.}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in \code{vname}.}

\item{add}{Collection to store assertion messages.}
}
\description{
Checkmate wrapper for argument checks. See \code{\link[checkmate]{makeAssertion}}
for more informations.
}
