% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{check_genotype_table}
\alias{check_genotype_table}
\title{Test if the matrix is a genotype_table of all the combinations possible of the
number of colums in a 0/1 format.}
\usage{
check_genotype_table(x)
}
\arguments{
\item{x}{The matrix to check.}
}
\value{
TRUE or an error message.
}
\description{
Test if the matrix is a genotype_table of all the combinations possible of the
number of colums in a 0/1 format.
}
