% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_magellan.R
\name{equalize_row_order}
\alias{equalize_row_order}
\title{Reorder the rows of a fitness landscape according to \code{genotype_table}.}
\usage{
equalize_row_order(genotype_table, fl)
}
\arguments{
\item{genotype_table}{A matrix of 0 and 1. The rows correspond to different genotypes
and the columns to the mutations that are considered for these genotypes.
A genotype (at a certain row) has a given mutation when there is a 1 in the
corresponding column. A row with only zeros correspond to the wild type.}

\item{fl}{A matrix of logical and a last colum of reals. A fitness landscape
with the first columns corresponding to a genotype_table and the last column
to the fitness of the genotype of each row.}
}
\value{
A fitness landscape similar to \code{fl} with the rows reordered according to
\code{genotype_table}.
}
\description{
Reorder the rows of a fitness landscape according to \code{genotype_table}.
}
