% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_magellan.R
\name{fl_generate_magellan}
\alias{fl_generate_magellan}
\title{Calls the fl_generate binary from MAGELLAN using system(\code{command}), and return
a fitness landscape generated by MAGELLAN reordered as \code{genotype_table}.}
\usage{
fl_generate_magellan(nb_mut, genotype_table, command)
}
\arguments{
\item{nb_mut}{A natural number. Number of mutations (or loci in MAGELLAN) used to
form the different genotypes from \code{genotype_table}.}

\item{genotype_table}{A matrix of 0 and 1. The rows correspond to different genotypes
and the columns to the mutations that are considered for these genotypes.
A genotype (at a certain row) has a given mutation when there is a 1 in the
corresponding column. A row with only zeros correspond to the wild type.}

\item{command}{A command with the file path to the binary fl_generate and argument
accepted by this binary.}
}
\value{
A fitness landscape with the \code{nb_mut} first columns corresponding
to the \code{genotype_table} and the \code{nb_mut}+1 column corresponding to
the fitness of the genotype of each row.
}
\description{
Calls the fl_generate binary from MAGELLAN using system(\code{command}), and return
a fitness landscape generated by MAGELLAN reordered as \code{genotype_table}.
}
