% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno-fitness_fgm_iso.R
\name{ftop_fgm_iso}
\alias{ftop_fgm_iso}
\title{Generates the phenotypic coordinates, in a \code{n}-dimension(s) phenotypic space,
corresponding to a given fitness in the FGM. Use the inverse of
the fitness function of an isotrope FGM (with parameters \code{n}, \code{maxfitness},
\code{alpha}, \code{Q}) to compute the euclidian distance to the phenotypic
optimum (\code{pheno_opt} -at the origin by default- at which the fitness =
\code{maxfitness}). This distance is reported on the first coordinate and all
the other \code{n}-1 coordinates are equal to the coordinates of the phenotypic
optimum.}
\usage{
ftop_fgm_iso(fitness, n, maxfitness, alpha = 1/2, Q = 2,
  pheno_opt = numeric(n))
}
\arguments{
\item{fitness}{A real number. The fitness of the phenotype. Must be lower or
equal to \code{maxfitness}.}

\item{n}{A natural number. Number of dimensions of the phenotypic space.}

\item{maxfitness}{A real number. The maximum fitness in the landscape. The
fitness at the phenotypic optimum (\code{pheno_opt}).}

\item{alpha}{A strictly positive real number. Scaling factor for the fitness
function. Default=1/2 in the cannonical FGM with a quadratic fitness function.}

\item{Q}{A strictly positive number. "Shape" of the fitness function. Default=2
in the cannonical FGM with a quadratic fitness function.}

\item{pheno_opt}{A vector of coordinates for the position of the phenotypic
optimum at which the fitness is equal to \code{maxfitness}. Its length must be
equal to \code{n}. Default = numeric(n).}
}
\value{
A vector of \code{n} coordinates in the phenotypic space.
}
\description{
Generates the phenotypic coordinates, in a \code{n}-dimension(s) phenotypic space,
corresponding to a given fitness in the FGM. Use the inverse of
the fitness function of an isotrope FGM (with parameters \code{n}, \code{maxfitness},
\code{alpha}, \code{Q}) to compute the euclidian distance to the phenotypic
optimum (\code{pheno_opt} -at the origin by default- at which the fitness =
\code{maxfitness}). This distance is reported on the first coordinate and all
the other \code{n}-1 coordinates are equal to the coordinates of the phenotypic
optimum.
}
\examples{
ftop_fgm_iso(fitness = 0, n = 3, maxfitness = 1)
ftop_fgm_iso(fitness = 0, n = 3, maxfitness = 1, alpha = 1/2, Q = 2, pheno_opt = c(1,1,1))
}
