% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coselected_mutation.R
\name{generate_coselected_mutation}
\alias{generate_coselected_mutation}
\title{Generates \code{nb_mut} co_selected mutations by calling sequentially
\code{\link{generate_selected_mutation}}. So that the second mutation happens
on the genetic background of the first selected mutants, the third mutations
on the genetic background of the second selected mutants...}
\usage{
generate_coselected_mutation(nb_mut, n, lambda, maxfitness, pheno_wt,
  alpha = 1/2, Q = 2, m = n, nb_mut_rand = 10^4)
}
\arguments{
\item{nb_mut}{A natural number. Number of mutations in the output.}

\item{n}{A natural number. Number of dimensions of the phenotypic space in
which the random mutations are drawn.}

\item{lambda}{A positive real number. Variance of the effect of mutations on
phenotype per phenotypic dimension.}

\item{maxfitness}{A real number. The maximum fitness in the landscape. The
fitness at the phenotypic optimum (\code{pheno_opt}).}

\item{pheno_wt}{A vector of real numbers. Phenotypic coordinates of the wild-type.
Its length must be equal to the number of columns of \code{pheno_mut_effect}.}

\item{alpha}{A strictly positive real number. Scaling factor for the fitness
function. Default=1/2 in the cannonical FGM with a quadratic fitness function.}

\item{Q}{A strictly positive number. "Shape" of the fitness function. Default=2
in the cannonical FGM with a quadratic fitness function.}

\item{m}{A natural number inferior to \code{n}. Level of restricted pleiotropy
which corresponds to the number of dimensions for which each mutation have a
non-zero phenotypic effect. i.e. effects are equal to 0 in n-m dimensions.
Default=\code{n} which corresponds to full pleiotropy.}

\item{nb_mut_rand}{A natural number superior or equal to \code{nb_mut}. Number
of random mutations to draw from which the mutant will be selected. This number
must be large compared to \code{nb_mut}. Default=10^4.}
}
\value{
A matrix of \code{nb_mut} rows of coselected mutations effects, ordered
by rows from the first mutation to the last mutation in increasing order. Note
that if no beneficial mutation is drawn for a given mutant, NA are introduced
from the corresponding row in the phenotype matrix to the end of the matrix.
In this case \code{nb_mut_rand} must be increased.
}
\description{
Generates \code{nb_mut} co_selected mutations by calling sequentially
\code{\link{generate_selected_mutation}}. So that the second mutation happens
on the genetic background of the first selected mutants, the third mutations
on the genetic background of the second selected mutants...
}
\examples{
generate_coselected_mutation(nb_mut = 5, n = 3, lambda = 0.1, maxfitness = 1,
                             pheno_wt = c(-1, 0, 0))
generate_coselected_mutation(nb_mut = 5, n = 3, lambda = 0.1, maxfitness = 1,
                             pheno_wt = c(-1, 0, 0), alpha = 1, Q = 0.5,
                             m = 2, nb_mut_rand = 10^5)
}
