% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_mutation.R
\name{generate_random_mutation}
\alias{generate_random_mutation}
\title{Draw \code{nb_mut} mutation(s) effect(s) on phenotype of \code{n} dimensions
with a variance of \code{lambda} per phenotypic dimension.
Also applies any restricted pleiotropy of level \code{m}}
\usage{
generate_random_mutation(nb_mut, n, lambda, m = n)
}
\arguments{
\item{nb_mut}{A natural number. Number of selected mutations in the output.}

\item{n}{A natural number. Number of dimensions of the phenotypic space in
which the random mutations are drawn.}

\item{lambda}{A positive real number. Variance of the effect of mutations on
phenotype per phenotypic dimension.}

\item{m}{A natural number inferior to \code{n}. Level of restricted pleiotropy
which corresponds to the number of dimensions for which each mutation have a
non-zero phenotypic effect. i.e. effects are equal to 0 in n-m dimensions.
Default=\code{n} which corresponds to full pleiotropy.}
}
\value{
A matrix with \code{nb_mut} rows of random mutation effects in \code{n}
columns (phenotypic dimension(s)).
}
\description{
Draw \code{nb_mut} mutation(s) effect(s) on phenotype of \code{n} dimensions
with a variance of \code{lambda} per phenotypic dimension.
Also applies any restricted pleiotropy of level \code{m}
}
\examples{
#full pleiotropy
generate_random_mutation(nb_mut = 3, n = 3, lambda = 0.1)
#restricted pleiotropy
generate_random_mutation(nb_mut = 3, n = 3, lambda = 0.1, m = 1)
}
