% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_magellan.R
\name{magellan}
\alias{magellan}
\alias{model_Fix}
\alias{model_Mult}
\alias{model_HoC}
\alias{model_NK}
\alias{model_Ising}
\alias{model_EggBox}
\alias{model_Optimum}
\title{Functions generating the fitness of the mutants' genotypes in \code{genotype_table}
The mutants are generated by a call to the fl_generate binary from MAGELLAN. See
\href{http://rstudio.com}{Rstudio}.
Warning : These functions return malthusian fitnesses (also called log fitnesses).
i.e. they use the argument "-L" in MAGELLAN.}
\usage{
model_Fix(nb_mut, genotype_table, fitness_wt, f, path2_fl_generate)

model_Mult(nb_mut, genotype_table, fitness_wt, s, S, d, path2_fl_generate)

model_HoC(nb_mut, genotype_table, fitness_wt, H, path2_fl_generate)

model_NK(nb_mut, genotype_table, fitness_wt, K, r, path2_fl_generate)

model_Ising(nb_mut, genotype_table, fitness_wt, i, I, c, path2_fl_generate)

model_EggBox(nb_mut, genotype_table, fitness_wt, e, E, path2_fl_generate)

model_Optimum(nb_mut, genotype_table, fitness_wt, o, O, p, P,
  path2_fl_generate)
}
\arguments{
\item{nb_mut}{A natural number. Number of mutations (or loci in MAGELLAN) used to
form the different genotypes from \code{genotype_table}.}

\item{genotype_table}{A matrix of 0 and 1. The rows correspond to different genotypes
and the columns to the mutations that are considered for these genotypes.
A genotype (at a certain row) has a given mutation when there is a 1 in the
corresponding column. A row with only zeros correspond to the wild type.}

\item{fitness_wt}{A real number. Fitness of the wild type used as reference.}

\item{f}{A real number. Add a fixed contribution}

\item{path2_fl_generate}{File path to the binary fl_generate of MAGELLAN.}

\item{s}{A real number. Mean fitness selection coefficient per locus (fitness
is then fitness_wt + s)}

\item{S}{A real number. Stdev for fitness for each locus (if 0 use fix fitness)}

\item{d}{A real number. Can set a diminishing (negative) or increasing (positive) return as you approach the peak}

\item{H}{A positive real number. Stdev for fitness for House-of-Cards (normal centered on 0)}

\item{K}{A positive integer. Kaufman NK with N=\code{nb_mut} and K<=N}

\item{r}{A logical value. If TRUE interacting loci are chosen at random. If FALSE
they are the neighbors.}

\item{i}{A real number. Mean cost for incompatibility.}

\item{I}{A positive real number. Stdev cost for incompatibility (if 0 use fix cost)}

\item{c}{A logical value. If TRUE last interacs with first (loci are arranged on a circle).}

\item{e}{A real number. Every other genotype is +/- e}

\item{E}{A positive real number. Add noise on the mean effect for eggbox}

\item{o}{A real number. Optimum to target.}

\item{O}{A positive real number. The spread around the optimal point, the larger
the less selection for optimum}

\item{p}{A real number. The mean production value for each non 0 allele.}

\item{P}{A positive real number. The associated stdev.}
}
\value{
A vector of real numbers. Each element is the fitness of the corresponding
random mutants in \code{genotype_table}. They are in the same order as the rows of
\code{genotype_table}.
}
\description{
Functions generating the fitness of the mutants' genotypes in \code{genotype_table}
The mutants are generated by a call to the fl_generate binary from MAGELLAN. See
\href{http://rstudio.com}{Rstudio}.
Warning : These functions return malthusian fitnesses (also called log fitnesses).
i.e. they use the argument "-L" in MAGELLAN.
}
