% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{model_fgmcsmut_2env}
\alias{model_fgmcsmut_2env}
\title{Function generating the fitness of the mutants' genotypes in \code{genotype_table}
in both the new and the reference environment. Mutants are produced by
combining \code{nb_mut} coselected mutations generated from the wt with fitness
\code{fitness_wt_ref} (in \code{fun_args}) in the reference environment.
Fitnesses are computed using an isotropic FGM in two dimensions with the parameters
passed to \code{fun_args}. See \code{\link{generate_coselected_mutation}}
for more information on the coselected mutations. The fitness of the mutants in the
new environment are then computed in a two (or one) dimensions isotropic FGM
with the dimensions corresponding to the two (or the unique) first dimensions
in which the mutants were generated in the environment of reference. For more
informations on the positionning of the optimum of the new environment see
\code{\link{pos_new_env}}. For more informations on the computation of fitness
from genotype see \code{\link{fitness_mutant_genotype}}.}
\usage{
model_fgmcsmut_2env(nb_mut, genotype_table, fitness_wt_new_env, lambda,
  maxfitness, alpha, Q, theta, fun_args)
}
\arguments{
\item{nb_mut}{A natural number. Number of mutations in the output.}

\item{genotype_table}{A matrix of 0 and 1. The rows correspond to different genotypes
and the columns to the mutations that are considered for these genotypes.
A genotype (at a certain row) has a given mutation when there is a 1 in the
corresponding column. A row with only zeros correspond to the wild type.}

\item{fitness_wt_new_env}{A real number. Fitness of the wild type in the new environment.}

\item{lambda}{A positive real number. Variance of the effect of mutations on
phenotype per phenotypic dimension.}

\item{maxfitness}{A real number. The maximum fitness in the landscape. The
fitness at the phenotypic optimum (\code{pheno_opt}).}

\item{alpha}{A strictly positive real number. Scaling factor for the fitness
function. Default=1/2 in the cannonical FGM with a quadratic fitness function.}

\item{Q}{A strictly positive number. "Shape" of the fitness function. Default=2
in the cannonical FGM with a quadratic fitness function.}

\item{theta}{A angle in radian between 0 and 2 * pi. Angle between the optimum
of the environment of reference (first dimension of \code{pheno_wt_ref}), the
\code{pheno_wt_ref} and the optimum of the new environement.}

\item{fun_args}{List of parameters used to generate coselected mutants in the reference environment.
Must take the form :
fun_args = list(fitness_wt_ref = #, n_ref = #, lambda_ref = #, maxfitness_ref = #,
alpha_ref = #, Q_ref = #, m_ref = #). Can optionnaly take the element nb_mut_ran.
For more information on these parameters, see the parameters in \code{\link{model_fgmcsmut}}}
}
\value{
A vector of 2 * \code{nb_mut} real numbers. The elements from 1 to \code{nb_mut}
are the fitnesses of the corresponding coselected mutants in \code{genotype_table}
(in the same order as the rows) in the reference environment. The elements from
\code{nb_mut} + 1 to 2 * \code{nb_mut} are the fitnesses of the same mutants
(in the same order) in the new environment.
}
\description{
Function generating the fitness of the mutants' genotypes in \code{genotype_table}
in both the new and the reference environment. Mutants are produced by
combining \code{nb_mut} coselected mutations generated from the wt with fitness
\code{fitness_wt_ref} (in \code{fun_args}) in the reference environment.
Fitnesses are computed using an isotropic FGM in two dimensions with the parameters
passed to \code{fun_args}. See \code{\link{generate_coselected_mutation}}
for more information on the coselected mutations. The fitness of the mutants in the
new environment are then computed in a two (or one) dimensions isotropic FGM
with the dimensions corresponding to the two (or the unique) first dimensions
in which the mutants were generated in the environment of reference. For more
informations on the positionning of the optimum of the new environment see
\code{\link{pos_new_env}}. For more informations on the computation of fitness
from genotype see \code{\link{fitness_mutant_genotype}}.
}
