% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{model_fgmsmut}
\alias{model_fgmsmut}
\title{Function generating the fitness of the mutants' genotypes in \code{genotype_table}
The mutants are produced by combining \code{nb_mut} selected mutations generated
from the wt with fitness \code{fitness_wt}. Fitnesses are computed using an
isotropic FGM with the parameter \code{n}, \code{lambda}, \code{maxfitness},
\code{alpha}, \code{Q}, \code{m}. See \code{\link{generate_selected_mutation}}
for more information on the selected mutations and \code{\link{fitness_mutant_genotype}}
for the fitness of the genotypes.}
\usage{
model_fgmsmut(nb_mut, genotype_table, fitness_wt, n, lambda, maxfitness,
  alpha, Q, m, fun_args)
}
\arguments{
\item{nb_mut}{A natural number. Number of mutations in the output.}

\item{genotype_table}{A matrix of 0 and 1. The rows correspond to different genotypes
and the columns to the mutations that are considered for these genotypes.
A genotype (at a certain row) has a given mutation when there is a 1 in the
corresponding column. A row with only zeros correspond to the wild type.}

\item{fitness_wt}{A real number. Fitness of the wild type in the new environment.}

\item{n}{A natural number. Number of dimensions of the phenotypic space in
which the random mutations are drawn.}

\item{lambda}{A positive real number. Variance of the effect of mutations on
phenotype per phenotypic dimension.}

\item{maxfitness}{A real number. The maximum fitness in the landscape. The
fitness at the phenotypic optimum (\code{pheno_opt}).}

\item{alpha}{A strictly positive real number. Scaling factor for the fitness
function. Default=1/2 in the cannonical FGM with a quadratic fitness function.}

\item{Q}{A strictly positive number. "Shape" of the fitness function. Default=2
in the cannonical FGM with a quadratic fitness function.}

\item{m}{A natural number inferior to \code{n}. Level of restricted pleiotropy
which corresponds to the number of dimensions for which each mutation have a
non-zero phenotypic effect. i.e. effects are equal to 0 in n-m dimensions.
Default=\code{n} which corresponds to full pleiotropy.}

\item{fun_args}{List with a single element called nb_mut_rand. For more information
see \code{\link{generate_selected_mutation}}}
}
\value{
A vector of real numbers. Each element is the fitness of the corresponding
selected mutants in \code{genotype_table}. They are in the same order as the rows of
\code{genotype_table}.
}
\description{
Function generating the fitness of the mutants' genotypes in \code{genotype_table}
The mutants are produced by combining \code{nb_mut} selected mutations generated
from the wt with fitness \code{fitness_wt}. Fitnesses are computed using an
isotropic FGM with the parameter \code{n}, \code{lambda}, \code{maxfitness},
\code{alpha}, \code{Q}, \code{m}. See \code{\link{generate_selected_mutation}}
for more information on the selected mutations and \code{\link{fitness_mutant_genotype}}
for the fitness of the genotypes.
}
