% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{pos_new_env}
\alias{pos_new_env}
\title{Returns the coordinates of a new optimum in the plane corresponding to the
first two dimensions of \code{pheno_wt_ref}. The position is computed using the
inverse of the fitness function of an isotrope FGM on the \code{fitness_wt_new_env}.}
\usage{
pos_new_env(pheno_wt_ref, fitness_wt_new_env, maxfitness, alpha = 1/2,
  Q = 2, theta)
}
\arguments{
\item{pheno_wt_ref}{A vector of real number. Phenotype of the wild type.}

\item{fitness_wt_new_env}{A real number. Fitness of the wild type in the new
env.}

\item{maxfitness}{A real number. The maximum fitness in the landscape. The
fitness at the phenotypic optimum (\code{pheno_opt}).}

\item{alpha}{A strictly positive real number. Scaling factor for the fitness
function. Default=1/2 in the cannonical FGM with a quadratic fitness function.}

\item{Q}{A strictly positive number. "Shape" of the fitness function. Default=2
in the cannonical FGM with a quadratic fitness function.}

\item{theta}{A angle in radian between 0 and 2 * pi. Angle between the optimum
of the environment of reference (first dimension of \code{pheno_wt_ref}), the
\code{pheno_wt_ref} and the optimum of the new environement.}
}
\value{
A pair of coordinates for the position of the optimum of the new environment
corresponding to the two first dimensions of \code{pheno_wt_ref}. If \code{pheno_wt_ref}
has a single dimension, \code{theta} is ignored and the position of the
optimum is given in this single dimension.
}
\description{
Returns the coordinates of a new optimum in the plane corresponding to the
first two dimensions of \code{pheno_wt_ref}. The position is computed using the
inverse of the fitness function of an isotrope FGM on the \code{fitness_wt_new_env}.
}
