% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_selected_mutation.R
\name{prob_esc_drift}
\alias{prob_esc_drift}
\title{Computes the probability of escaping drift for a mutation in a clonal population,
using its selection coefficient. For more info see Gerrish et al.(1998) Appendix I}
\usage{
prob_esc_drift(sel_coeff)
}
\arguments{
\item{sel_coeff}{A vector of real numbers corresponfing to selectioncoefficients.}
}
\description{
Computes the probability of escaping drift for a mutation in a clonal population,
using its selection coefficient. For more info see Gerrish et al.(1998) Appendix I
}
