% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptof_fgm_iso.R
\name{ptof_fgm_iso}
\alias{ptof_fgm_iso}
\title{Generates the fitness corresponding to the phenotypic coordinates of \code{phenotype}
using Fisher's Geometric Model (FGM, Fisher (1930)). Use the fitness function
of an isotrope FGM (with parameters \code{n},  \code{maxfitness}, \code{alpha},
\code{Q}) to compute the fitness (malthusian fitness) of phenotype(s)
\eqn{log(W) = \code(maxfitness)- \code{alpha} * \code{phenotype}^(\code{Q}/2)}
(see Tenaillon et al. (2007))
See also \code{\link{ftop_fgm_iso}} for inverse function.}
\usage{
ptof_fgm_iso(phenotype, maxfitness, alpha = 1/2, Q = 2,
  pheno_opt = numeric(dim(phenotype)[2]))
}
\arguments{
\item{phenotype}{A vector, matrix of real number(s) (phenotypic
coordinate(s)). For a matrix or a data.frame, the rows are phenotypes and the
columns phenotypic dimensions.}

\item{maxfitness}{A real number. The maximum fitness in the landscape. The
fitness at the phenotypic optimum (\code{pheno_opt}).}

\item{alpha}{A strictly positive real number. Scaling factor for the fitness
function. Default=1/2 in the cannonical FGM with a quadratic fitness function.}

\item{Q}{A strictly positive number. "Shape" of the fitness function. Default=2
in the cannonical FGM with a quadratic fitness function.}

\item{pheno_opt}{A vector of coordinates for the position of the phenotypic
optimum at which the fitness is equal to \code{maxfitness}. Its length must be
equal to the the number of column of phenotype.}
}
\value{
A vector of fitnesses of length equal to the number of phenotype(s) (row(s)) in \code{phenotype}
}
\description{
Generates the fitness corresponding to the phenotypic coordinates of \code{phenotype}
using Fisher's Geometric Model (FGM, Fisher (1930)). Use the fitness function
of an isotrope FGM (with parameters \code{n},  \code{maxfitness}, \code{alpha},
\code{Q}) to compute the fitness (malthusian fitness) of phenotype(s)
\eqn{log(W) = \code(maxfitness)- \code{alpha} * \code{phenotype}^(\code{Q}/2)}
(see Tenaillon et al. (2007))
See also \code{\link{ftop_fgm_iso}} for inverse function.
}
\examples{
#' @examples
ptof_fgm_iso(phenotype = matrix(1:9, 3, 3, byrow = TRUE), maxfitness = 1)
ptof_fgm_iso(phenotype = matrix(1:9, 3, 3, byrow = TRUE), maxfitness = 1,
alpha = 1/2, Q = 2, pheno_opt = c(1,1,1))
}
