% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_mutation.R
\name{random_mutation}
\alias{random_mutation}
\title{Draw \code{nb_mut} mutation(s) effect(s) on phenotype of \code{n} dimensions}
\usage{
random_mutation(n, nb_mut, lambda)
}
\arguments{
\item{n}{A natural number. Number of dimensions of the multivariate normal
distribution from which the phenotypic effects of mutations are drawn}

\item{nb_mut}{A natural number. Number of mutations for each of which a phenotypic
effect is drawn.}

\item{lambda}{A strictly positive real number. Variance of mutation effects on
phenotype}
}
\value{
A matrix of \code{nb_mut} rows and \code{n} columns.
}
\description{
Draw \code{nb_mut} mutation(s) effect(s) on phenotype of \code{n} dimensions
}
