% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_magellan.R
\name{scale_to_wt_fitness}
\alias{scale_to_wt_fitness}
\title{Function generating the fitness of the mutants' genotypes in \code{genotype_table}
The mutants are produced by combining \code{nb_mut} selected mutations generated
from the wt with fitness \code{fitness_wt}. Fitnesses are computed using an
isotropic FGM with the parameter \code{n}, \code{lambda}, \code{maxfitness},
\code{alpha}, \code{Q}, \code{m}. See \code{\link{generate_selected_mutation}}
for more information on the selected mutations and \code{\link{fitness_mutant_genotype}}
for the fitness of the genotypes.}
\usage{
scale_to_wt_fitness(nb_mut, genotype_table, fitness_wt, fl)
}
\arguments{
\item{nb_mut}{A natural number. Number of mutations (or loci in MAGELLAN) used to
form the different genotypes from \code{genotype_table}.}

\item{genotype_table}{A matrix of 0 and 1. The rows correspond to different genotypes
and the columns to the mutations that are considered for these genotypes.
A genotype (at a certain row) has a given mutation when there is a 1 in the
corresponding column. A row with only zeros correspond to the wild type.}

\item{fitness_wt}{A real number. Fitness of the wild type used as reference.}

\item{fl}{A matrix of logical and a last colum of reals. A fitness landscape
with the first columns corresponding to a genotype_table and the last column
to the fitness of the genotype of each row.}
}
\value{
A vector of fitnesses with the fisrt element equal to \code{fitness_wt}
and the other scaled to correspond to mutants with a wild type with fitness
\code{fitness_wt}.
}
\description{
Calls the fl_generate binary from MAGELLAN using system(\code{command}), and return
a fitness landscape generated by MAGELLAN reordered as \code{genotype_table}.
}
