% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fl.R
\name{simulate_fl}
\alias{simulate_fl}
\title{Apply \code{simulation_model} to the rows of the matrix \code{parameter} either
sequentially or in parallel (depends on \code{ncore}).}
\usage{
simulate_fl(parameter, simulation_model, ncore = 1, file_output = NULL,
  multi_file = TRUE, ...)
}
\arguments{
\item{parameter}{A matrix of type numeric. Each row correspond to the arguments
in \code{simulation_model}. the order of the column must be the same as the arguments
in \code{simulation_model}.}

\item{simulation_model}{A function or a function name. The function's arguments
must cotrrespond to a row of \code{parameter} with same order.}

\item{ncore}{A natural number. Number of cores used for the simulations. If
\code{ncore} = 1 simulation_model is applied sequentially over the row of
\code{parameter}. If \code{ncore} > 1 simulation_model is applied in parallel
over the row of \code{parameter}. The parallelization is not avilable for "Windows".}

\item{file_output}{Full path to the output file for the simulations. If NULL
no output file is created.}

\item{multi_file}{Logical. Applies only if \code{file_output} is not NULL. If
TRUE the results of the simulations are saved in two separated files with the
suffixes "-simulation" and "-parameter". If FALSE parameters and simulations
are saved in the same file with first columns for the parameters and the last
for the corresponding simulations.}

\item{...}{Extra arguments which will be passed to \code{write.table} when an
output is saved in a file. e.g. \code{sep}}
}
\value{
A list of three elements. The first element is the matrix of \code{parameter}.
If some parameters are unnamed, all the parameters are renamed to P1, P2...
The second element is the matrix of simulations, in which each row is the result
of \code{simulation_model} applied on the corresponding row in \code{parameter}. The
columns of this matrix are named S1, S2...
The third element is the output of \code{system.time} for the total simulation
process.
}
\description{
Apply \code{simulation_model} to the rows of the matrix \code{parameter} either
sequentially or in parallel (depends on \code{ncore}).
}
\examples{
simulation_model <- function(m, sd) {rnorm(n = 10, mean = m, sd = sd)}
parameter <- cbind(1:10, seq(0.1, 1, 0.1))
simulate_fl(parameter, simulation_model, ncore = 1)
}
