% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_mutation.R
\name{random_mutation}
\alias{random_mutation}
\title{Draw \code{nb_mut} mutation(s) effect(s) on phenotype of \code{n} dimensions}
\usage{
random_mutation(n, nb_mut, lambda)
}
\arguments{
\item{n}{A natural number. Number of dimensions of the phenotypic space in
which the random mutations are drawn.}

\item{nb_mut}{A natural number. Number of mutations in the output.}

\item{lambda}{A positive real number. Variance of the effect of mutations on
phenotype per phenotypic dimension.}
}
\value{
A matrix of \code{nb_mut} rows and \code{n} columns.
}
\description{
Draw \code{nb_mut} mutation(s) effect(s) on phenotype of \code{n} dimensions
}
