% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_mutation.R
\name{restricted_pleiotropy}
\alias{restricted_pleiotropy}
\title{Creates restricted pleiotropy (as defined in Chevin et al.(2010)) of a level
\code{m} to a matrix of phenotypic effects of mutations. Insert a 0 at \code{m}
random positions for each row of the matrix.}
\usage{
restricted_pleiotropy(mutation, m)
}
\arguments{
\item{mutation}{A matrix of real number. The rows are mutations and the columns
phenotypic dimensions.}

\item{m}{A natural number inferior to the number of colums of \code{mut}.
Number of dimensions for which each mutation have a non-zero phenotypic effect.}
}
\value{
A matrix of the same number of rows and columns as \code{mut} with
\code{m} random positions equal to zero in each rows
}
\description{
Creates restricted pleiotropy (as defined in Chevin et al.(2010)) of a level
\code{m} to a matrix of phenotypic effects of mutations. Insert a 0 at \code{m}
random positions for each row of the matrix.
}
