% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fl.R
\name{simulate_fl}
\alias{simulate_fl}
\title{Apply \code{simulation_model} to the rows of the matrix \code{parameter} either
sequentially or in parallel (depends on \code{ncore}).}
\usage{
simulate_fl(parameter, simulation_model, empirical_fl = NULL,
  ncore = 1, file_output = NULL, multi_file = TRUE,
  fun_args = list(), output_args = list(), ...)
}
\arguments{
\item{parameter}{A matrix of type numeric. Each row correspond to the arguments
in \code{simulation_model}. the order of the column must be the same as the arguments
in \code{simulation_model}.}

\item{simulation_model}{A function or a character designing the \code{model_type}
from \code{\link{generate_model}}. If a function is provided, its arguments must
correspond to a row of \code{parameter} with same order.}

\item{empirical_fl}{A matrix corresponding to an empirical fitness landscape.
The first ncol-1 columns correspond to a genotype table. The rows correspond
to different genotypes and the columns to the mutations that are considered for
these genotypes. A genotype (at a certain row) has a given mutation when there
is a 1 in the corresponding column.
The column to of empirical_fl correspond to the fitness of each genotype. Only
the fitness of the wt (i.e. the row in the genotype table with only zeros) is
used.}

\item{ncore}{A natural number. Number of cores used for the simulations. If
\code{ncore} = 1 simulation_model is applied sequentially over the row of
\code{parameter}. If \code{ncore} > 1 simulation_model is applied in parallel
over the row of \code{parameter}. The parallelization is not avilable for "Windows".}

\item{file_output}{Full path to the output file for the simulations. If NULL
no output file is created.}

\item{multi_file}{Logical. Applies only if \code{file_output} is not NULL. If
TRUE the results of the simulations are saved in two separated files with the
suffixes "-simulation" and "-parameter". If FALSE parameters and simulations
are saved in the same file with first columns for the parameters and the last
colums for the corresponding simulations.}

\item{fun_args}{List of optionnal argument for certain \code{model_type}:
\describe{
  \item{nb_mut_rand}{for "fgmsmut" and "fgmcsmut". See \code{\link{generate_selected_mutation}} and \code{\link{generate_coselected_mutation}}}
}}

\item{output_args}{List of arguments which will be passed to \code{\link[utils]{write.table}}
if \code{file_output} = TRUE.}

\item{...}{Extra arguments which will be passed to \code{\link[utils]{read.table}} if an empirical
fitness landscape is provided as a file in \code{empirical_fl}}
}
\value{
A list of three elements. The first element is the matrix of \code{parameter}.
If some parameters are unnamed, all the parameters are renamed to P1, P2...
The second element is the matrix of simulations, in which each row is the result
of \code{simulation_model} applied on the corresponding row in \code{parameter}. The
columns of this matrix are named S1, S2...
The third element is the output of \code{system.time} for the total simulation
process.
}
\description{
Apply \code{simulation_model} to the rows of the matrix \code{parameter} either
sequentially or in parallel (depends on \code{ncore}).
}
\examples{
simulation_model <- function(m, sd) {rnorm(n = 10, mean = m, sd = sd)}
parameter <- cbind(1:10, seq(0.1, 1, 0.1))
simulate_fl(parameter, simulation_model, ncore = 1)
}
